结构优化，层次分明

+++ seckill-web/src/main/java/com/plusl/web/interceptor/RequireLogin.java
@@ -13,7 +13,7 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 @Target(METHOD)
 public @interface RequireLogin {
 
-    int seconds();
+    long seconds();
 
     int maxCount();
